refactor: 폴더 구조 일부 변경
