+++ src/assets/index.ts
@@ -1,0 +1,4 @@
+export { default as Facebook } from "./svg/Facebook";
+export { default as Instagram } from "./svg/Instagram";
+export { default as Twitter } from "./svg/Twitter";
+export { default as Youtube } from "./svg/Youtube";
