+++ src/assets/svg/CardImageEmptyCaseIcon.tsx
