--- src/assets/svg/FooterIcons/index.ts
@@ -1,4 +1,0 @@
-export { default as Facebook } from "./Facebook";
-export { default as Instagram } from "./Instagram";
-export { default as Twitter } from "./Twitter";
-export { default as Youtube } from "./Youtube";
